Fix RNG offset desyncs by setting it once (#101)
Initially, the RNG offset is set to the system time using `os.time()`.

When a new Slippi connection is established, the seed will be set to the
the Slippi RNG offset the first time a match starts.

Every time a new stage is loaded, the seed is advanced deterministically
by XORing it with the prime number 397.

This should keep clients and the host in sync throughout the session.